Make per-session JSON stores crash-safe with shared atomic writes
Extracts the ui-state temp-file+rename+queue logic into
atomic-file-write.ts and reuses it in JsonFileStore, so attachment
files can no longer be truncated by a crash mid-write. Adds
JsonFileStore.remove for upcoming orphan cleanup.
